requirements.txt
paho-mqtt
playsound
pgi
